move over all queue instrumentation to the instrumentation queue module
